modifiedcodeloginusingcommoncomponent
